The demo drives the new controls, and its vault is nobody's
Four things, and the second one turned out to be the reason for the first.

THE STORYBOARD GAINS A CAMERA ACT AND A DATE ACT. Three new verbs, since the demo
could only click and hover:

  {drag: [dx, dy], target}   press, glide, release -- with the button HELD, which
                             is not moveTo plus click: the ribbon's brush and the
                             disc's pan both read `buttons` on every move, and a
                             press and a release with button-up moves between them
                             is a press and a release with nothing between them.
  {wheel: n, target}         n notches, one at a time with a beat between. Sigma
                             animates each over zoomDuration, so firing them back
                             to back shows one blur instead of the steps.
  {dblclick, target}         two clicks inside the double-click window.

And two synthetic target kinds, because the camera and the ribbon's handles are
positions rather than elements: ["stage", "centre"] or ["stage", "0.3,0.4"], and
["brush", "from" | "to" | "window"]. The brush kinds resolve from the same
geometry the control draws with, so a beat aims where the handle IS rather than
where the storyboard guessed it would be.

39 beats, all landing: zoom in, pan, double-click to reset, zoom again, the corner
button, then both range handles and the band window dragged independently.

THE DEMO VAULT IS NO LONGER A MIRROR OF ANYONE'S. It walked a real vault and
reproduced its shape under invented names, which meant the output depended on
whose machine it ran on -- and, worse, it inherited that vault's DATE SHAPE. One
real vault is 86% of its notes in a single year with a thin tail of ten more:
honest, and useless for demonstrating a date-range control, because the ribbon is
one spike and nine years of near-empty strip.

So it is a preset over make-test-vault.mjs, which already carried a full declared
spec, rather than a second generator -- two generators for one job drift. 1400
notes over two dense years, every month populated, ramping toward the present.

THE MIRRORING IS NOT GONE, it moved to make-mirror-vault.mjs, because CONTRIBUTING
points at it for a good reason: if the layout misbehaves on your vault, the shape
IS the bug report, and a generic fixture cannot reproduce a shape it does not
have. Passing --vault to make-demo-vault now fails loudly and says where to go.

Three more top-level folders, taking the fixture to eighteen -- eight past the ten
colour slots, so the neutral fallback carries a third of the disc rather than a
corner of it.

TWO BUGS FOUND BY LOOKING AT THE RESULT:

  The recency lean was INVERTED. `DAYS - pow(rnd(), 0.55) * 365` leans a
  distance-BACK toward larger, so the burst landed at the beginning of its own
  twelve months and the newest weeks came out emptiest -- a heatmap whose
  right-hand edge, the part that is today, was the sparsest thing on it. The
  exponent has to be above 1 here and below 1 in the tail, which is not symmetry;
  it is the opposite, because one is a distance back from the end and the other a
  distance forward from the start.

  The ribbon bars were a SOLID SLAB on a two-year vault. sqrt against the maximum
  was right for eleven years, where the busiest month holds 631 against a median
  under 40 -- and did the opposite damage here, where the ratio is three: every
  bar came out between 68% and 100%. Linear against p90 (floored at a third of the
  max) reads in both. A real spike clips, which is honest, and costs one month's
  precision to give every other month some.

AND THE WINDOW DRAG IS LIVE. The brush still previews and lands on release, because
dragging it re-lays out the whole disc; the window only re-buckets the band, which
is a different budget and can afford it. rebuildBand() computes the quantised week
first and skips heatBuild entirely when it has not moved -- so the pill glides and
the grid steps by weeks, which is not a compromise but the truth about the two
things. Measured on the 10k fixture: median 17ms, p90 17ms, one 33ms frame.

  node scripts/smoke.mjs --vault ./demo-vault        # 30/30
  node scripts/smoke.mjs --vault ./test-vault-10y    # 30/30
  node scripts/demo.mjs                              # 39/39 beats
